RSA Update V2
This is a pretty major update with a brand-new Key File Storage system. Here is the list of some new features:

- New Key Storage system allows for easily storing and loading key files for easier transmissions.
- Many, many optimizations such as typo fixing, changing Strings to StringBuilders instead of conjugating them, and removing redundant functions and variables.
- Made the encryption process very slightly faster. (any speed increase is good) 

I will also try to add some kind of documentation and an example of using KeyFiles.
NOTE: This version is 100% compatible with the older RSA version and I always try to make newer versions backwards compatible.